+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-13 18:43:44
-// Commit: fix(state): correct layout constraints and overflow in state
+// Updated at: 2026-05-14 11:46:34
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '1115';
-  final String lastMessage = 'fix(state): correct layout constraints and overflow in state';
-  final DateTime timestamp = DateTime.parse('2026-05-13T18:43:44+05:00');
+  final String commitRef = '1133';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-05-14T11:46:34+05:00');
 }